requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.66.1
pandas>=2.0.3
pydantic>=2.5.2